refactor: boldText 파일 분리
