+++ mission/src/components/index.tsx
@@ -1,21 +1,7 @@
 import { ChangeEvent, useState } from "react";
 import searchIcon from "../assets/search.svg";
 import { dummy } from "../data.ts";
-
-const boldText = (text: string, query: string) => {
-  if (!query.trim()) return text;
-
-  const parts = text.split(new RegExp(`(${query})`, "gi"));
-  return parts.map((part, index) =>
-    query.toLowerCase() === part.toLowerCase() ? (
-      <strong key={index} className="font-bold">
-        {part}
-      </strong>
-    ) : (
-      part
-    )
-  );
-};
+import boldText from "../utils/boldText.tsx";
 
 const Input = () => {
   const [text, setText] = useState("");
